[DOP-11980] - add pydantic smart union
