+++ onetl/impl/frozen_model.py
@@ -19,4 +19,5 @@ from onetl.impl.base_model import BaseModel
 
 class FrozenModel(BaseModel):
     class Config:
+        smart_union = True
         frozen = True
